feat(vxm): Launch of development features without font-end

+++ vx_library/manager/ShellManager.py
@@ -1,4 +1,6 @@
+import os, sys, signal, time
 from .utils import use_sudo
+from .logger import Logger
 
 
 class ShellManager:
@@ -26,23 +28,49 @@ class ShellManager:
         if not feature_name:
             return
 
-        vite_process = get_vite_process(directory)
-        if not vite_process:
-            return
+        if os.path.exists(f"{directory}/package.json"):
+            vite_process = get_vite_process(directory)
+            if not vite_process:
+                ShellRequests.unload_feature(feature_name)
+                return
 
-        vite_process.start()
+            vite_process.start()
 
-        if ShellRequests.start_feature(feature_name):
-            print(f"  \033[92m➜\033[0m  Vixen: start feature '{feature_name}'")
-            vite_process.join()
+            if ShellRequests.start_feature(feature_name):
+                print(f"  \033[92m➜\033[0m  Vixen: start feature '{feature_name}'")
+                vite_process.join()
+            else:
+                print(
+                    f"  \033[31m➜\033[0m  Vixen: Error on starting feature '{feature_name}'"
+                )
+                vite_process.terminate()
+
+            if ShellRequests.ping():
+                ShellRequests.unload_feature(feature_name)
+
         else:
-            print(
-                f"  \033[31m➜\033[0m  Vixen: Error on starting feature '{feature_name}'"
-            )
-            vite_process.terminate()
 
-        if ShellRequests.ping():
-            ShellRequests.unload_feature(feature_name)
+            def signal_handler(sig, frame):
+                Logger.log("Stop feature '{feature_name}' [Dev mode]")
+
+                if ShellRequests.ping():
+                    ShellRequests.unload_feature(feature_name)
+
+                sys.exit(0)
+
+            signal.signal(signal.SIGINT, signal_handler)
+
+            Logger.log("Vixen feature '{feature_name}' [Dev mode]")
+
+            if ShellRequests.start_feature(feature_name):
+                Logger.log("start feature '{feature_name}'")
+            else:
+                Logger.log("Error on starting feature '{feature_name}'", "ERROR")
+
+            Logger.log("Press Ctrl+C to exit...")
+
+            while True:
+                time.sleep(1)
 
     @staticmethod
     @use_sudo(False)
